added more robust error handling to spo and more unit tests
