+++ tests/test_class_initialization.py
