--- tests/todo.txt
@@ -1,6 +1,0 @@
-- change in gamma
-- create and test multiple portfolios
-
-after pull
-- delete branch
-- create new branch
